+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Recommendation: for ease of reading, use the following order:
 - GQL: Added new `webhookSubscription::rotateSecret()` method to replace current secret
 ### Changed
 - Webhook subscription secrets are now stored encrypted in the database
+### Fixed
+- Flow process state projections now directly handle flow scope removal events
 
 ## [0.251.0] - 2025-10-16
 ### Added